src/main/java/com/flyonsky/weixin/data/material/CreateImageReceive.java
package com.flyonsky.weixin.data.material;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.flyonsky.weixin.data.ServiceReceive;

/**
 * 上传图文消息中的图片，但并不作为永久的图片素材响应
 * @author Administrator
 *
 */
@JsonInclude(JsonInclude.Include.NON_NULL)
public class CreateImageReceive extends ServiceReceive{

	@JsonProperty("url")
	private String  url;

	public String getUrl() {
		return url;
	}

	public void setUrl(String url) {
		this.url = url;
	}
}
